Added remaining page
On to do page:
task cards no longer expand to show options.
On focus-schedule page:
when box is checked settings for task shows, on large screens animation for estimated time should be moved over to the right and settings remain on the left.
Once all chosen tasks have been approved by user, and focus schedule is clicked, the page goes to focused which shows the breakdown and timer of the schedule.